03_custom_gpts/additional_material/02_gpt_actions/03_onboarding_gpt/action-code/onboarding-service/app/core/security.py
from datetime import datetime, timedelta
from typing import Any
from jose import jwt

from app import settings


ALGORITHM: str = "HS256"


def create_access_token(subject: str | Any, expires_delta: timedelta) -> str:
    expire = datetime.utcnow() + expires_delta
    to_encode = {"exp": expire, "sub": str(subject)}
    encoded_jwt = jwt.encode(
        to_encode, settings.SECRET_KEY, algorithm=ALGORITHM)
    return encoded_jwt
